+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 0.2.0 - 2018-06-10
 - Add `Manager.AddKey` method for more intuitive key creation
 - Add `Hotkey.Name` attribute for better reference in output
-- Use `time.After` in `SeekHotkeyLoop` to sleep safely
+- Use `time.After` in `Listen` to sleep safely
+- Add `KeyPress` logic to parse actual time of press from `WindowsMessage`
 
 ## 0.1.0 - 2018-06-08
 - Working prototype with five-second alerts for each major map spawn